moved out verbatim db
