+++ src/corpus/__init__.py
@@ -3,10 +3,10 @@ import tempfile
 import urllib.request
 import logging
 from subprocess import check_call
-from arbiter.verbatimdb.__main__ import generate_db
+from arbiter.verbatimdb.db import generate_db
 
 MALICIOUS_BOOTSTRAP_URL = os.getenv('MALICIOUS_BOOTSTRAP_URL')
-ARCHIVE_PW = os.getenv('ARCHIVE_PASSWORD')
+ARCHIVE_PASSWORD = os.getenv('ARCHIVE_PASSWORD')
 
 
 class DownloadToFileSystemCorpus(object):
@@ -21,27 +21,27 @@ class DownloadToFileSystemCorpus(object):
         return os.path.join(self.base_dir, "malicious")
 
     @property
-    def benign_pth(self):
-        return os.path.join(self.base_dir, "malicious")
+    def benign_path(self):
+        return os.path.join(self.base_dir, "benign")
 
     def download_and_unpack(self):
         logging.info("Downloading to {0}".format(self.base_dir))
         for is_mal, package_name in [(True, "malicious.tgz.gpg"), (False, "benign.tgz.gpg")]:
 
-            mal_dir = self.mal_path if is_mal else self.benign_pth
-            archive_pth = os.path.join(mal_dir, package_name)
+            mal_dir = self.mal_path if is_mal else self.benign_path
+            archive_path = os.path.join(mal_dir, package_name)
             archive_dst = mal_dir
             if not os.path.isdir(archive_dst):
-                os.mkdir(archive_dst)
-            archive_tgz = archive_pth.rstrip(".gpg")
-            r = urllib.request.urlretrieve("{0}/{1}".format(self.url, package_name), archive_pth)
-            check_call(["gpg", "--batch", "--passphrase", ARCHIVE_PW, "--decrypt",
+                os.makedirs(archive_dst, exist_ok=True)
+            archive_tgz = archive_path.rstrip(".gpg")
+            r = urllib.request.urlretrieve("{0}/{1}".format(self.url, package_name), archive_path)
+            check_call(["gpg", "--batch", "--passphrase", ARCHIVE_PASSWORD, "--decrypt",
                         "--no-use-agent", "--cipher-algo", "AES256",
-                         "--yes", "-o", archive_tgz, archive_pth])
+                         "--yes", "-o", archive_tgz, archive_path])
             # can do with python tar file, but being lazy
             check_call(["tar", "xf", archive_tgz, "-C", archive_dst])
             os.unlink(archive_tgz)
-            os.unlink(archive_pth)
+            os.unlink(archive_path)
 
     def _get_pth_listing(self, p):
         artifacts = []
@@ -54,12 +54,9 @@ class DownloadToFileSystemCorpus(object):
         return self._get_pth_listing(self.mal_path)
 
     def get_benign_file_list(self):
-        return self._get_pth_listing(self.benign_pth)
+        return self._get_pth_listing(self.benign_path)
 
     def download_truth(self):
         u = "{0}/{1}".format(self.url, self.truth_fname)
         logging.info("Fetching truth database {0}".format(u))
         r = urllib.request.urlretrieve(u, self.truth_db_pth)
-
-    def generate_truth(self):
-        d = generate_db(self.truth_db_pth, self.mal_path, self.benign_pth)
